pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fruit-segmentation"
version = "0.1.0"
description = "Hybrid U-Net segmentation pipeline for automated fruit sorting using ConvNeXt-V2 and Swin-V2 backbones."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "timm>=0.9.16",
    "numpy>=1.26.0",
    "pillow>=10.2.0",
    "pyyaml>=6.0.1",
    "matplotlib>=3.8.0",
    "pandas>=2.2.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.4.0",
    "albumentations>=1.3.1",
    "opencv-python-headless>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
]
notebook = [
    "ipykernel>=6.29.0",
    "ipywidgets>=8.1.0",
]

[project.scripts]
train = "train.train:main"
infer = "inference.inference:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.hatch.build.targets.wheel]
packages = ["models", "utils", "dataset", "train", "inference", "visualization"]
